requirements.txt
langserve
fastapi
uvicorn
openai 
langchain
langchain-openai
langchain_pinecone
langchain[docarray]
langchain-community
docarray
pydantic==1.10.8
pytubefix 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
sse_starlette
nest_asyncio
streamlit
git+https://github.com/openai/whisper.git